Working on the calculator logic
It includes a function to obtain the button type, a function to perform the required operation that receives both numbers and the operator as input parameters, and a function to carry out all the calculator logic, thus leaving the code clean and scalable, and only calling these functions in the event. It also includes a function to display what is needed on the calculator screen.